refactor: move code into src/ and fix imports
