+++ eslint.config.mjs
@@ -33,6 +33,9 @@ const typeScriptRuleset = merge(...typescript, {
 		'@typescript-eslint/explicit-member-accessibility': ['error', { accessibility: 'no-public' }],
 		'@typescript-eslint/array-type': ['error', { default: 'array-simple' }],
 		'@typescript-eslint/consistent-type-definitions': ['error', 'type'],
+
+		// those aren't absolute paths; ref tsconfig .compilerOptions.paths
+		'import-x/no-absolute-path': 'off',
 	},
 	settings: {
 		'import/resolver': {
