+++ src/pages/favicon.ico
