+++ src/pages/guide/images/farmAfterSendingCBsWithTimer.png
