+++ src/pages/guide/images/farmReadyToReceiveCBs.png
