+++ src/pages/guide/images/newContractIcon.jpg
