+++ tsconfig.json
@@ -1,5 +1,10 @@
 {
 	"compilerOptions": {
+		"rootDir": ".",
+		"baseUrl": ".",
+		"paths": {
+			"@/*": ["src/*"]
+		},
 		"target": "esnext",
 		"lib": ["dom", "dom.iterable", "esnext"],
 		"allowJs": true,
@@ -17,11 +22,7 @@
 			{
 				"name": "next"
 			}
-		],
-		"paths": {
-			"@/*": ["./*"],
-			"/*": ["./*"]
-		}
+		]
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
 	"exclude": ["node_modules"]
